# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Envelope_PaperExpandedEntity_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PaperExpandedEntity"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaInfo"
          },
          "success": {
            "const": true,
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[PaperExpandedEntity]",
        "type": "object"
      },
...
      "Envelope_Union_PaperExpandedEntity__PaperEntity__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaperExpandedEntity"
              },
              {
                "$ref": "#/components/schemas/PaperEntity"
              }
            ],
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaInfo"
          },
          "success": {
            "const": true,
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[Union[PaperExpandedEntity, PaperEntity]]",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaInfo"
          },
          "success": {
            "const": false,
            "default": false,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "error",
          "meta"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "LayoutDensity": {
        "enum": [
          "auto",
          "normal",
          "compact",
          "dense"
        ],
        "title": "LayoutDensity",
        "type": "string"
      },
...
      "PaperCreate": {
        "additionalProperties": false,
        "properties": {
          "duration": {
            "exclusiveMinimum": 0.0,
            "title": "Duration",
            "type": "integer"
          },
          "ownerId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ownerid"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionRef"
            },
            "title": "Questions",
            "type": "array"
          },
          "subject": {
            "minLength": 1,
            "title": "Subject",
            "type": "string"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "totalMarks": {
            "exclusiveMinimum": 0.0,
            "title": "Totalmarks",
            "type": "integer"
          }
        },
        "required": [
          "title",
          "subject",
          "duration",
          "totalMarks"
        ],
        "title": "PaperCreate",
        "type": "object"
      },
...
      "QuestionOrder": {
        "enum": [
          "paper",
          "categorized"
        ],
        "title": "QuestionOrder",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "opaque",
        "description": "Native-client credential injection point. Device-session issuance is owned by CLE-18.",
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "description": "HttpOnly browser session cookie.",
        "in": "cookie",
        "name": "testpapers_session",
        "type": "apiKey"
      },
      "csrfToken": {
        "description": "Required with cookie authentication for unsafe HTTP methods.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "The canonical HTTP contract for `/api/v1`. Browser clients use Cookie + CSRF authentication; native clients inject Bearer credentials. WebSocket messages are described by `x-testpapers-websocket`.",
    "summary": "Versioned Cloud contract shared by TestPapers Web, Desktop, and Mobile.",
    "title": "TestPaper Backend",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/papers": {
      "post": {
        "operationId": "create_paper",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaperCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PaperExpandedEntity_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication is required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The caller is not allowed to perform this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The write rate limit was exceeded."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "cookieAuth": [],
            "csrfToken": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create Paper",
        "tags": [
          "papers"
        ]
      }
    },
    "/api/v1/papers/draft-download": {
      "post": {
        "operationId": "download_draft_paper",
        "parameters": [
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "docx",
              "pattern": "^(docx|tex)$",
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaperDraftDownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Generated document bytes.",
            "headers": {
              "Content-Disposition": {
                "description": "RFC 6266 attachment filename.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Cloud-Draft-Export": {
                "description": "Present and set to `true` for a shared cloud-draft export.",
                "schema": {
                  "enum": [
                    "true"
                  ],
                  "type": "string"
                }
              },
              "X-Draft-Export": {
                "description": "Present and set to `true` for an unsaved paper-draft export.",
                "schema": {
                  "enum": [
                    "true"
                  ],
                  "type": "string"
                }
              },
              "X-Export-Format": {
                "description": "Resolved export format.",
                "schema": {
                  "enum": [
                    "docx"
                  ],
                  "type": "string"
                }
              },
              "X-Layout-Density": {
                "description": "Resolved document layout density.",
                "schema": {
                  "enum": [
                    "auto",
                    "normal",
                    "compact",
                    "dense"
                  ],
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication is required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The caller is not allowed to perform this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The write rate limit was exceeded."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "cookieAuth": [],
            "csrfToken": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Download Draft Paper",
        "tags": [
          "papers"
        ]
      }
    },
    "/api/v1/papers/generate": {
      "post": {
        "operationId": "generate_paper",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaperGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication is required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The caller is not allowed to perform this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The write rate limit was exceeded."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "cookieAuth": [],
            "csrfToken": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Generate Paper",
        "tags": [
          "papers"
        ]
      }
    },
    "/api/v1/papers/{paper_public_id}": {
      "get": {
        "operationId": "get_paper",
        "parameters": [
          {
            "in": "path",
            "name": "paper_public_id",
            "required": true,
            "schema": {
              "title": "Paper Public Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expand",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expand"
            }
          },
          {
            "in": "query",
            "name": "includeAnswer",
            "required": false,
            "schema": {
              "default": true,
              "title": "Includeanswer",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Union_PaperExpandedEntity__PaperEntity__"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication is required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The caller is not allowed to perform this operation."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested resource was not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get Paper",
        "tags": [
          "papers"
        ]
      },
      "patch": {
        "operationId": "update_paper",
        "parameters": [
          {
            "in": "path",
            "name": "paper_public_id",
            "required": true,
            "schema": {
              "title": "Paper Public Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaperUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PaperEntity_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication is required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The caller is not allowed to perform this operation."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested resource was not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The write rate limit was exceeded."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "cookieAuth": [],
            "csrfToken": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update Paper",
        "tags": [
          "papers"
        ]
      }
    },
    "/api/v1/papers/{paper_public_id}/download": {
      "get": {
        "operationId": "download_paper",
        "parameters": [
          {
            "in": "path",
            "name": "paper_public_id",
            "required": true,
            "schema": {
              "title": "Paper Public Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "docx",
              "pattern": "^(docx|tex)$",
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "questionOrder",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/QuestionOrder",
              "default": "paper"
            }
          },
          {
            "in": "query",
            "name": "includeAnswer",
            "required": false,
            "schema": {
              "default": true,
              "title": "Includeanswer",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "layoutDensity",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/LayoutDensity",
              "default": "auto"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Generated document bytes.",
            "headers": {
              "Content-Disposition": {
                "description": "RFC 6266 attachment filename.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Cloud-Draft-Export": {
                "description": "Present and set to `true` for a shared cloud-draft export.",
                "schema": {
                  "enum": [
                    "true"
                  ],
                  "type": "string"
                }
              },
              "X-Draft-Export": {
                "description": "Present and set to `true` for an unsaved paper-draft export.",
                "schema": {
                  "enum": [
                    "true"
                  ],
                  "type": "string"
                }
              },
              "X-Export-Format": {
                "description": "Resolved export format.",
                "schema": {
                  "enum": [
                    "docx"
                  ],
                  "type": "string"
                }
              },
              "X-Layout-Density": {
                "description": "Resolved document layout density.",
                "schema": {
                  "enum": [
                    "auto",
                    "normal",
                    "compact",
                    "dense"
                  ],
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication is required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The caller is not allowed to perform this operation."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested resource was not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Download Paper",
        "tags": [
          "papers"
        ]
      }
    },
...
  },
...
}